Write or Remove Attribute Table in a GeoPackage

gpkg_write_attributes(): Specify a target geopackage and name for new table. For adding attributes, specify the new data as data.frame. The table name will be registered in the gpkg_contents table. Optionally include a custom description and/or use a template object to define the spatial extent associated with attribute data.

gpkg_remove_attributes(): Remove an attribute table and corresponding gpkg_contents record

gpkg_write_attributes(
  x,
  table,
  table_name,
  description = "",
  template = NULL,
  overwrite = FALSE,
  append = FALSE
)

gpkg_remove_attributes(x, table_name)

Arguments

x

A geopackage object

table

A data.frame

table_name

character. The name for table in x

description

Optional description. Default ""

template

A list (containing elements "ext" and "crs", or a terra object. These objects defining xmin/ymin/xmax/ymax and spatial reference system for the attribute table.

overwrite

Overwrite? Default FALSE

append

Append? Default FALSE

Value

logical. TRUE on successful table write or remove.