Create an empty feature table

Create an empty feature table and associated entries for gpkg_spatial_ref_sys, and gpkg_geometry_columns.

gpkg_create_empty_features(
  x,
  table_name,
  column_name = "geom",
  geometry_type_name = "GEOMETRY",
  srs_id = 4326,
  z = 0L,
  m = 0L,
  contents = TRUE,
  description = "",
  ext = c(-180, -90, 180, 90)
)

Arguments

x

A geopackage Object

table_name

character. New table name.

column_name

character. Geometry column name. Default "geom"

geometry_type_name

character. Geometry type name. Default: "GEOMETRY"

srs_id

integer. Spatial Reference System ID. Must be defined in gpkg_spatial_ref_sys table.

z

integer. Default: 0

m

integer. Default: 0

contents

logical. If TRUE (default) add the new table to gpkg_contents table.

description

character. Description for gpkg_contents table. Default: ""

ext

numeric. A numeric vector of length four specifying the bounding box extent.

Value

integer result of gpkg_execute(). Returns 1 if a new geometry record is appended to gpkg_geometry_columns table.